changed uint8_t n to uint32_t n

+++ part2/part2Main.c
@@ -7,7 +7,7 @@
 #include "hitRate.h"
 
 int main() {
-	uint8_t n;
+	uint32_t n;
 	uint32_t blockDataSize;
 	uint32_t totalDataSize;
 	uint32_t address;
